Update PHTML files
Each PHTML file corresponding to the relevant integration, e.g.
edit.phtml has been edited so that it now contains the hosted link for
the integration.

At some point it may be worth customizing this so that a customer could
use their own hosted link that's configured by adjusting the
SITE_VARIABLES in their store.